Split MethodCallWithArgsParentheses
